% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDpack.R
\name{ALDpack}
\alias{ALDpack}
\alias{rald_mix}
\alias{dald_mix}
\alias{pald_mix}
\alias{qald_mix}
\title{Asymmetric Laplace distribution}
\usage{
rald_mix(n, mu = 0, sigma = 1, p = 0.5, plot = FALSE)

dald_mix(x, mu = 0, sigma = 1, p = 0.5)

pald_mix(q, mu = 0, sigma = 1, p = 0.5, lower.tail = TRUE)

qald_mix(prob, mu = 0, sigma = 1, p = 0.5, lower.tail = TRUE)
}
\arguments{
\item{n}{: number of observations}

\item{mu}{: location parameter}

\item{sigma}{: scale parameter}

\item{p, prob}{: probability at which to calculate quantile}

\item{plot}{: creates a basic plot of AL(mu,sigma,p)}

\item{x, q}{: quantiles}

\item{lower.tail}{: decides b/w P(X<=p) or P(X>p) for p/q}
}
\value{
\itemize{
\item {\code{rald_mix}} {returns a vector of random numbers from ALD(mu,sigma,p).}
\item {\code{dald_mix}} {returns returns density of AL(mu,sigma,p) at point x.}
\item {\code{pald_mix}} {returns CDF prob of AL(mu,sigma,p) at quantile q.}
\item {\code{qald_mix}} {returns inverse CDF quantile of AL(mu,sigma,p) at prob.}
}
}
\description{
Cumulative density, Probability distribution function, Quantile
function and Random generation for the Asymmetric Laplace distribution with
quantile p, location parameter mu and scale parameter sigma.
}
\details{
The Asymmetric Laplace distribution (ALD), which has the following pdf:

\deqn{f(x;\mu,\sigma,p) = \frac{p(1-p)}{\sigma} exp{-\frac{(x-\mu)}{\sigma} (p-I(x \le \mu))}}{f(x;\mu,\sigma,p) = p(1-p)/\sigma exp{-(x-\mu)/\sigma (p-I(x \le \mu))}}

If not specified, p=0.5, mu = 0, sigma = 1.
}
\examples{
rald_mix(1e5,mu=5,sigma=10,p=0.5,plot=TRUE)
dald_mix(c(4,5),mu=0,sigma=1,p=0.5)
pald_mix(c(1,4),mu=0,sigma=1,p=0.5,lower.tail=TRUE)
qald_mix(0.5,mu=0,sigma=1,p=0.5,lower.tail=TRUE)

}
\references{
Keming Yu & Jin Zhang (2005) A Three-Parameter Asymmetric Laplace Distribution
and Its Extension, Communications in Statistics - Theory and Methods, 34:9-10,
1867-1879, DOI: 10.1080/03610920500199018

Kobayashi, Genya. (2011). Gibbs Sampling Methods for Bayesian Quantile Regression.
J Stat Comput Simul. 81. 1565. 10.1080/00949655.2010.496117.
}
